+++ src/main/java/com/airbnl/managerservice/service/Interfaces/IReservationService.java
@@ -5,7 +5,7 @@ import com.airbnl.managerservice.model.Reservation;
 import java.util.List;
 
 public interface IReservationService {
-    Reservation save(Reservation reservation, int managerId);
-    List<Reservation> getAllByHotelAndManagerId(int hotelId, int managerId);
-    Reservation getByReservationIdAndManagerId(int reservationId, int managerId);
+    Reservation save(Reservation reservation, Long managerId);
+    List<Reservation> getAllByHotelAndManagerId(Long hotelId, Long managerId);
+    Reservation getByReservationIdAndManagerId(Long reservationId, Long managerId);
 }
